api_design_env/server/requirements.txt
openenv-core>=0.2.2
fastapi>=0.104.0
uvicorn>=0.24.0
